Upload challenge1
This challenge is very simple. You just need to create a server that responds with "Maratona Full Cycle 2.0" when you access localhost:3000.

This project also need to be transformed in a Docker Image so that when you run

```sudo docker container run -p 3000:3000 --name challenge1 axell13/fullcycle-challenge1```

you can test it in any machine that has Docker.